Fix numeric type checks; remove unused fields
Remove several unused VearchConfig properties and clean up numeric type handling across multiple servlets and metadata classes. Replace duplicated Double checks with proper Float/float checks in ToolMetadata, AnnotationApiServlet, RouteServlet, and BankManagerByApiRouter; also adjust default numeric value logic. Remove get_current_date from the demo agent's tool list. Trim unused dependencies from oxygent-infra/pom.xml (commons-lang3, lombok, okhttp entries) as part of dependency cleanup.

+++ oxygent-core/src/main/java/com/jd/oxygent/core/oxygent/mcpservers/engine/metadata/ToolMetadata.java
@@ -89,7 +89,7 @@ public class ToolMetadata {
         if (type == Integer.class || type == int.class ||
                 type == Long.class || type == long.class ||
                 type == Double.class || type == double.class ||
-                type == Double.class || type == float.class) return "number";
+                type == Float.class || type == float.class) return "number";
         if (type == Boolean.class || type == boolean.class) return "boolean";
         return "string";
     }
